tavern: add more authentication tests
There were some mistakes in mapping the rust types to PostgreSQL types,
so many files were touched to correct those. The database script has
also been modified to use a schema, which makes it much simpler to drop
all tables, types, and whatever else from the database for testing
purposes.

For more on rust -> pgsql type mappings, see
https://docs.rs/sqlx/0.3.5/sqlx/postgres/types/index.html

+++ tavern_pathfinder/src/class.rs
@@ -28,8 +28,8 @@ pub struct Class {
     description: String,
     hit_die: String,
     starting_wealth: String,
-    bab_per_level: f32,
-    skills_per_level: i32,
+    bab_per_level: f64,
+    skills_per_level: i16,
     skills_attr: Attribute,
 }
